Refactored required classes into headers for client
