+++ inc/FtlTypes.h
