+++ inc/IConnectionTransport.h
