+++ test/functional/FunctionalTests.cpp
@@ -1,0 +1,8 @@
+/**
+ * @file FunctionalTests.cpp
+ * @author Hayden McAfee (hayden@outlook.com)
+ * @date 2020-11-25
+ * @copyright Copyright (c) 2020 Hayden McAfee
+ */
+
+#include <catch2/catch.hpp>
